refactor: simplificar contentview extraindo componentes reutilizaveis e reorganizando features
