+++ CodexPlus.xcodeproj/xcuserdata/ana.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<Bucket
+   uuid = "147971BC-20EE-45FA-8C13-E1AF6E9106C7"
+   type = "1"
+   version = "2.0">
+</Bucket>
