--- CodexPlus/DataSourceMock/DataSourceMock.swift
@@ -1,15 +1,0 @@
-//
-//  DataSourceMock.swift
-//  CodexPlus
-//
-//  Created by ana on 08/09/25.
-//
-
-import Foundation
-
-struct DataSourceMock {
-    let articles: [NewsArticle] = [
-        
-    ]
-}
-
