+++ CodexPlus/FeedView/AgronegociosView.swift
@@ -23,16 +23,11 @@ struct AgronegociosView: View {
                     ScrollView(.vertical, showsIndicators: false) {
                         LazyVStack(spacing: 16) {
                             ForEach(feedItems, id: \.idValue) { item in
-                                if let link = item.content?.url,
-                                   let url = URL(string: link.replacingOccurrences(of: "http://", with: "https://")) {
-                                    Link(destination: url) {
-                                        ModelFeedItem(item: item)
-                                    }
-                                    .buttonStyle(.plain)
-                                }
+                                NewsRowView(item: item)
                             }
                         }
                         .padding(.vertical, 8)
+                        
                     }
                     .refreshable {
                         await loadFeed()
@@ -42,6 +37,7 @@ struct AgronegociosView: View {
             }
             
         }
+        
         .task { await loadFeed() }
     }
